--- src/data/tailwind_classes.rs
@@ -1,25 +1,0 @@
-//! Tailwind CSS classes, used programmatically throughout the project.
-
-/// Colors for element backgrounds, like player names.
-pub static BG_COLORS: [&str; 4] = [
-    "bg-rose-400",
-    "bg-cyan-400",
-    "bg-green-400",
-    "bg-violet-400",
-];
-
-/// Colors for element borders, like the individual scores.
-pub static BORDER_COLORS: [&str; 4] = [
-    "border-rose-400",
-    "border-cyan-400",
-    "border-green-400",
-    "border-violet-400",
-];
-
-/// Colors for input elements, used to color the caret in the score input.
-pub static CARET_COLORS: [&str; 4] = [
-    "caret-rose-400",
-    "caret-cyan-400",
-    "caret-green-400",
-    "caret-violet-400",
-];
