+++ src/frontend/mod.rs
@@ -90,3 +90,24 @@ fn decorative_spheres(cx: Scope) -> Element {
         }
     ))
 }
+
+pub static BG_COLORS: [&str; 4] = [
+    "bg-rose-400",
+    "bg-cyan-400",
+    "bg-green-400",
+    "bg-violet-400",
+];
+
+pub static BORDER_COLORS: [&str; 4] = [
+    "border-rose-400",
+    "border-cyan-400",
+    "border-green-400",
+    "border-violet-400",
+];
+
+pub static CARET_COLORS: [&str; 4] = [
+    "caret-rose-400",
+    "caret-cyan-400",
+    "caret-green-400",
+    "caret-violet-400",
+];
